It works now
Biggest problems in this project:
1. Tried to implement random names, years, selected
elements. Failied to do so. Reverted back to not
using random but forgot to revert the element
selection back, so it was trying to access random
elements, which were usually larger than what was
available.
2. While writing the movies to file, intented to
write the title with size titleLen, but
accidentally wrote the title with size
`sizeof(titleLen)`, meaning it would only write 4
chars of the title, and fail to read the rest of
the movies.