4/10 Progress Report -- Took 5 hours
Set out to add an admin class that would be able to create, edit, and
delete products. Menus work but need refining in some places. Admin
methods do not work yet, but the structure is in place. Admin is a
friend class to StoreFront, so that it can access store's Inventory.
Admin must be constructed with a reference to the StoreFront object
being used. This is not a working program :[